constant files & switched to codewrapper component
