+++ src/routes/svelte-basics/bindings/+page.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import CodeBlock from '$lib/components/ui/code-block/code-block.svelte';
+	import CodeWrapper from '$lib/components/ui/code-wrapper/code-wrapper.svelte';
+	import { CODE, HEADER } from './constants';
 </script>
 
 <h1>Bindings</h1>
@@ -16,15 +17,7 @@
 	To create a binding on a text input element, you can use the <code>bind:value</code> directive. This
 	will create a two-way binding between the input element and a variable in your component.
 </p>
-
-<CodeBlock>
-	{`<script>
-				let name;
-			<\/script>
-
-				<input bind:value={name} />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.TEXT_INPUTS} />
 
 <p>
 	This means that not only will changes to the value of <code>name</code> update the input value,
@@ -38,15 +31,7 @@
 	the value of the input is always a number since Svelte takes care of the type conversion for you.
 </p>
 
-<CodeBlock>
-	{`<script>
-				let age;
-			<\/script>
-
-				<input type="number" bind:value={age} min="0" max="10" />
-				<input type="range" bind:value={age} min="0" max="10" />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.NUMERIC_INPUTS} />
 
 <h2>Checkbox inputs</h2>
 
@@ -55,31 +40,13 @@
 	binding between the checkbox input and a variable in your component.
 </p>
 
-<CodeBlock>
-	{`<script>
-				let isChecked;
-			<\/script>
-
-				<input type="checkbox" bind:checked={isChecked} />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.CHECKBOXES} />
 
 <h2>Select bindings</h2>
 
 <p>We can also use <code>bind:value</code> with <span class="font-bold">select</span> elements.</p>
 
-<CodeBlock>
-	{`<script>
-				let selectedOption;
-			<\/script>
-
-				<select bind:value={selectedOption}>
-					<option value="1">Option 1</option>
-					<option value="2">Option 2</option>
-					<option value="3">Option 3</option>
-				</select>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.SELECT} />
 
 <h2>Select multiple</h2>
 
@@ -88,18 +55,7 @@
 	selected will populate an array rather than selecting a single value.
 </p>
 
-<CodeBlock>
-	{`<script>
-				let selectedOptions = [];
-			<\/script>
-
-				<select multiple bind:value={selectedOptions}>
-					<option value="1">Option 1</option>
-					<option value="2">Option 2</option>
-					<option value="3">Option 3</option>
-				</select>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.MULTI_SELECT} />
 
 <p>
 	Press and hold the <kbd>Ctrl</kbd> key or the <kbd>command</kbd> key on MacOS to select multiple options.
@@ -116,22 +72,7 @@
 	inputs in the same group form an array of selected values.
 </p>
 
-<CodeBlock>
-	{`<script>
-				let radioOption = 1;
-				let checkboxOptions = [];
-			<\/script>
-
-				<input type="radio" value="1" bind:group={radioOption} />
-				<input type="radio" value="2" bind:group={radioOption} />
-				<input type="radio" value="3" bind:group={radioOption} />
-
-				<input type="checkbox" value="1" bind:group={checkboxOptions} />
-				<input type="checkbox" value="2" bind:group={checkboxOptions} />
-				<input type="checkbox" value="3" bind:group={checkboxOptions} />
-
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.GROUP} />
 
 <p>
 	Now, when you select a radio option, the value of <code>radioOption</code> will be updated. When
@@ -145,25 +86,14 @@
 	<code>bind:value</code>.
 </p>
 
-<CodeBlock>
-	{`<script>
-				let value;
-			<\/script>
-
-				<textarea bind:value={value}></textarea>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.TEXTAREA} />
 
 <p>
 	A really cool feature of Svelte is that, since the name of the variable is equal to the name of
 	the input property (value), you can use the shorthand syntax:
 </p>
 
-<CodeBlock>
-	{`
-				<textarea bind:value />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.BINDINGS} code={CODE.SHORTHAND} />
 
 <p>
 	<span class="underline">This trick applies to all bindings, not just textareas.</span>
