+++ src/routes/svelte-basics/events/+page.svelte
@@ -1,34 +1,6 @@
 <script lang="ts">
-	import CodeBlock from '$lib/components/ui/code-block/code-block.svelte';
-	const mouseEventsList = [
-		{ event: 'on:mouseover', description: 'Fires when the mouse pointer moves over an element.' },
-		{ event: 'on:mouseout', description: 'Fires when the mouse pointer moves out of an element.' },
-		{ event: 'on:mouseenter', description: 'Fires when the mouse pointer enters an element.' },
-		{ event: 'on:mouseleave', description: 'Fires when the mouse pointer leaves an element.' }
-	];
-
-	const keyboardEventsList = [
-		{ event: 'on:keydown', description: 'Fires when a key is pressed down.' },
-		{ event: 'on:keyup', description: 'Fires when a key is released.' },
-		{ event: 'on:keypress', description: 'Fires when a key is pressed and released.' }
-	];
-
-	const eventModifiersList = [
-		{ modifier: '|preventDefault', description: 'Prevents the default behavior of the event.' },
-		{ modifier: '|stopPropagation', description: 'Stops the event from propagating.' },
-		{
-			modifier: '|self',
-			description: 'Only triggers the event if the event was dispatched from the element itself.'
-		},
-		{ modifier: '|once', description: 'Only triggers the event once.' },
-		{ modifier: '|capture', description: 'Use capture phase instead of bubbling phase.' },
-		{ modifier: '|passive', description: 'Use passive event listener.' },
-		{ modifier: '|nonpassive', description: 'Use non-passive event listener.' },
-		{
-			modifier: '|trusted',
-			description: 'Only trigger the event if it was dispatched by a user action.'
-		}
-	];
+	import CodeWrapper from '$lib/components/ui/code-wrapper/code-wrapper.svelte';
+	import { CODE, EVENTS_LIST, HEADER } from './constants';
 </script>
 
 <h1>Events</h1>
@@ -44,9 +16,7 @@
 	The most common event in Svelte is the click event. It's used to listen for clicks on an element.
 </p>
 
-<CodeBlock>
-	{`<button on:click={handleClick}>Click me</button>`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.CLICK} />
 
 <p>
 	In the example above, the <code>on:click</code> directive is used to listen for a click event on
@@ -55,14 +25,14 @@
 
 <h2>Mouse events</h2>
 <ul>
-	{#each mouseEventsList as { event, description }}
+	{#each EVENTS_LIST.MOUSE as { event, description }}
 		<li><code>{event}</code> - {description}</li>
 	{/each}
 </ul>
 
 <h2>Keyboard events</h2>
 <ul>
-	{#each keyboardEventsList as { event, description }}
+	{#each EVENTS_LIST.KEYBOARD as { event, description }}
 		<li><code>{event}</code> - {description}</li>
 	{/each}
 </ul>
@@ -76,9 +46,7 @@
 	this:
 </p>
 
-<CodeBlock>
-	{`<button on:mouseover={() => {console.log('Mouse is hovering me')}}>Hover me</button>`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.INLINE_HANDLERS} />
 
 <h2>Event modifiers</h2>
 <p>
@@ -87,12 +55,11 @@
 	event handler. In the following example, we use the <code>|preventDefault</code> modifier to prevent
 	the default behavior of the click event:
 </p>
-<CodeBlock>
-	{`<button on:click|preventDefault={handleSubmit}>Submit</button>`}
-</CodeBlock>
+
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.MODIFIERS} />
 
 <ul>
-	{#each eventModifiersList as { modifier, description }}
+	{#each EVENTS_LIST.MODIFIERS as { modifier, description }}
 		<li><code>{modifier}</code> - {description}</li>
 	{/each}
 </ul>
@@ -110,36 +77,12 @@
 </p>
 
 <p>On our <span class="font-bold">parent</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import Child from './Child.svelte';
 
-				function handleMessage(e) {
-					console.log(e.detail.text)
-					};
-			<\/script>
-
-				<Child on:message={handleMessage} />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.COMPONENTS_PARENT} />
 
 <p>On our <span class="font-bold">child</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import { createEventDispatcher } from 'svelte';
 
-				const dispatch = createEventDispatcher();
-
-				function sayHello() {
-					dispatch('message', { text: 'Hello from child component!' });
-				};
-			<\/script>
-
-			<button on:click={sayHello}>
-				Click to say hello
-			</button>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.COMPONENTS_CHILD} />
 
 <p>
 	In the example above, the child component dispatches a <code>:message</code> event with the text
@@ -161,46 +104,13 @@
 </p>
 
 <p>On our <span class="font-bold">parent</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import Child from './Child.svelte';
-
-				function handleMessage(e) {
-					console.log(e.detail.text)
-					};
-			<\/script>
-
-				<Child on:message={handleMessage} />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.FORWARDING_PARENT} />
 
 <p>On our <span class="font-bold">child</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import GrandChild from './GrandChild.svelte';
-			<\/script>
-
-			<GrandChild on:message />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.FORWARDING_CHILD} />
 
 <p>On our <span class="font-bold">grandchild</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import { createEventDispatcher } from 'svelte';
-
-				const dispatch = createEventDispatcher();
-
-				function sayHello() {
-					dispatch('message', { text: 'Hello from grandchild component!' });
-				};
-			<\/script>
-
-			<button on:click={sayHello}>
-				Click to say hello
-			</button>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.FORWARDING_GRANDCHILD} />
 
 <p>
 	In the example above, an <code>on:message</code> event directive is passed to the
@@ -220,24 +130,7 @@
 </p>
 
 <p>On our <span class="font-bold">parent</span> component:</p>
-<CodeBlock>
-	{`<script>
-				import Child from './Child.svelte';
-
-				function handleClick(e) {
-					console.log('Clicked on child component');
-				}
-			<\/script>
-
-			<Child on:click={handleClick} />
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.DOM_PARENT} />
 
 <p>On our <span class="font-bold">child</span> component:</p>
-<CodeBlock>
-	{`
-			<button on:click>
-				Click me
-			</button>
-	`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.EVENTS} code={CODE.DOM_CHILD} />
