+++ src/routes/svelte-basics/logic/+page.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
-	import Button from '$lib/components/ui/button/button.svelte';
-	import Card from '$lib/components/ui/card/card.svelte';
-	import CodeBlock from '$lib/components/ui/code-block/code-block.svelte';
-	import { Code } from 'lucide-svelte';
+	import CodeWrapper from '$lib/components/ui/code-wrapper/code-wrapper.svelte';
+	import { CODE, HEADER } from './constants';
 </script>
 
 <h1>Logic</h1>
@@ -11,45 +9,19 @@
 	block types that can be used for common operations.
 </p>
 
-<h2>If blocks</h2>
+<h2>If and else blocks</h2>
 <p>
 	An if block is used to conditionally render content. It can be used with an else block to render
 	different content based on a condition.
 </p>
-<CodeBlock>
-	{`{#if condition}
-    <p>Render this content if the condition is true</p>
-{:else}
-    <p>Render this content if the condition is false</p>
-{/if}`}
-</CodeBlock>
 
-<h2>Else blocks</h2>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.IF_ELSE} />
 
-<p>
-	An else block is used in conjunction with an if block to render different content based on a
-	condition.
-</p>
-
-<CodeBlock>
-	{`{#if condition}
-    <p>Render this content if the condition is true</p>
-{:else}
-    <p>Render this content if the condition is false</p>
-{/if}`}
-</CodeBlock>
+<h2>Else if blocks</h2>
 
 <p>It's also possible to use else-if:</p>
 
-<CodeBlock>
-	{`{#if condition1}
-    <p>Render this content if condition1 is true</p>
-{:else if condition2}
-    <p>Render this content if condition2 is true</p>
-{:else}
-    <p>Render this content if neither condition1 nor condition2 is true</p>
-{/if}`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.ELSE_IF} />
 
 <h2>Block rules</h2>
 <p>An easy way to remember this syntax is as follows:</p>
@@ -66,19 +38,11 @@
 	similar to how you'd map over an array in React.
 </p>
 
-<CodeBlock>
-	{`{#each array as item}
-	<p>{item}</p>
-{/each}`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.EACH} />
 
 <p>Also like mapping in react, you can get the iteration number as the second param like so:</p>
 
-<CodeBlock>
-	{`{#each array as item, i}
-	<p>{i}: {item}</p>
-{/each}`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.KEYED_EACH} />
 
 <h2>Keyed each blocks</h2>
 
@@ -88,11 +52,7 @@
 	like so:
 </p>
 
-<CodeBlock>
-	{`{#each array as item (item.id)}
-	<p>{item}</p>
-{/each}`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.KEYED_EACH_2} />
 
 <h2>Await blocks</h2>
 
@@ -101,15 +61,7 @@
 	for fetching data from an API or performing other asynchronous operations.
 </p>
 
-<CodeBlock>
-	{`{#await promise}
-	<p>Loading...</p>
-{:then data}
-	<p>{data}</p>
-{:catch error}
-	<p>{error}</p>
-{/await}`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.LOGIC} code={CODE.AWAIT} />
 
 <p>
 	The first block ("Loading...") is displayed while the promise is resolving, which is a nice and
