+++ src/routes/svelte-basics/props/+page.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
 	import PropsExample from '$lib/components/examples/props-example-component/props-example.svelte';
 	import PropsMultipleExample from '$lib/components/examples/props-multiple-example-component /props-multiple-example.svelte';
-	import CodeBlock from '$lib/components/ui/code-block/code-block.svelte';
+	import CodeWrapper from '$lib/components/ui/code-wrapper/code-wrapper.svelte';
+	import { COMPONENT_CODE, HEADER } from './constants';
 </script>
 
 <h1>Props</h1>
@@ -11,23 +12,8 @@
 	declared in the child component's script tag, and can be used in the component's HTML and JS.
 </p>
 <p>Here is an example of a component that accepts props:</p>
-<CodeBlock>
-	{`<script>
-    export let exampleName = 'Props Example';
-<\/script>
 
-<div>
-    This is an example of a component that accepts props. The prop passed to this component is: <strong>{exampleName}</strong>
-</div>
-
-<style>
-    div {
-        background-color: #1f4a65;
-        padding: 1rem;
-        border-radius: 5px;
-    }
-</style>`}
-</CodeBlock>
+<CodeWrapper headerText={HEADER.PROPS} code={COMPONENT_CODE.PROPS} />
 
 <p>And here's the component that renders:</p>
 <PropsExample exampleName="Props Example" />
@@ -38,32 +24,20 @@
 </p>
 
 <p>If we have multiple props, we can pass them like so:</p>
-<CodeBlock>
-	{`<PropsMultipleExample name="Diogo" age="33" country="Portugal" city="Barreiro" />`}
-</CodeBlock>
+
+<CodeWrapper headerText={HEADER.PROPS} code={COMPONENT_CODE.MULTIPLE_PROPS} />
 
 <p>And here's the component that renders:</p>
 <PropsMultipleExample name="Diogo" age="33" country="Portugal" city="Barreiro" />
 
 <p>But since Svelte provides us with a shorthand syntax, we can also pass props like this:</p>
-<CodeBlock>
-	{`<PropsMultipleExample {name} {age} {country} {city} />`}
-</CodeBlock>
+
+<CodeWrapper headerText={HEADER.PROPS} code={COMPONENT_CODE.MULTIPLE_PROPS_SHORTHAND} />
 
 <p>Or we could even spread an object into the component as props:</p>
 
-<CodeBlock>
-	{`<script>
-    let props = {
-        name: 'Diogo',
-        age: 33,
-        country: 'Portugal',
-        city: 'Barreiro'
-    }
-<\/script>
+<CodeWrapper headerText={HEADER.PROPS} code={COMPONENT_CODE.SPREAD_PROPS} />
 
-<PropsMultipleExample {...props} />`}
-</CodeBlock>
 <br />
 <p>
 	Another nifty trick that Svelte allows is to access props that were not exported/declared on the
